% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ats.R
\name{.get_ats_df}
\alias{.get_ats_df}
\title{Extract Acoustic Trawl Survey Data}
\usage{
.get_ats_df(M, biomass = TRUE)
}
\arguments{
\item{M}{A list of  model outputs. Each item should contain at least
the elements \code{yr_ats}, \code{ob_ats}, \code{eb_ats}, \code{sd_ob_ats}, \code{ot_ats}, \code{et_ats}, and \code{sd_ot_ats}.}

\item{biomass}{A logical flag. If \code{TRUE}, data extracted is for biomass;
if \code{FALSE}, data for another metric is extracted. Default is \code{TRUE}.}
}
\value{
A dataframe with columns: year, Model, obs, pre, lb, and ub.
}
\description{
Extracts data for either biomass or some other metric based on the given  models in the list \code{M}.
The function formats the data into a consistent structure for further processing.
}
