% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-avo.R
\name{.get_avo_df}
\alias{.get_avo_df}
\title{Get avo data into frame}
\usage{
.get_avo_df(M)
}
\arguments{
\item{M}{A list of models where each model has the needed attributes.}

\item{biomass}{Logical indicating if biomass data should be used. If FALSE, other measurements are used.}
}
\value{
A data frame with columns for year, model, observed values, predicted values, lower bounds, and upper bounds.
}
\description{
This function iterates through a list of models and extracts year and associated measurements.
}
