% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bts.R
\name{.get_bts_df}
\alias{.get_bts_df}
\title{Extract bottom trawl survey data from a list of models}
\usage{
.get_bts_df(M, biomass = TRUE)
}
\arguments{
\item{M}{A list of models where each model has the needed attributes.}

\item{biomass}{Logical indicating if biomass data should be used. If FALSE, other measurements are used.}
}
\value{
A data frame with columns for year, model, observed values, predicted values, lower bounds, and upper bounds.
}
\description{
This function iterates through a list of models and extracts year and associated measurements.
}
