% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cope.R
\name{.get_cope_df}
\alias{.get_cope_df}
\title{Extract copE data}
\usage{
.get_cope_df(M)
}
\arguments{
\item{M}{list object created by read_admb function}
}
\value{
dataframe of spawning biomass
}
\description{
Extract copE data
}
