% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cpue.R
\name{.get_cpue_df}
\alias{.get_cpue_df}
\title{Extract CPUE data}
\usage{
.get_cpue_df(M)
}
\arguments{
\item{M}{list object created by read_admb function}
}
\value{
dataframe of observed and predicted CPUE
}
\description{
Extract CPUE data
}
