% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mnage.R
\name{.get_mnage_df}
\alias{.get_mnage_df}
\title{Extract mean age from pollock model run}
\usage{
.get_mnage_df(M)
}
\arguments{
\item{M}{list object created by read_admb function}
}
\value{
dataframe of spawning biomass
}
\description{
Extract mean age from pollock model run
}
