% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-Nage3.R
\name{.get_Nage_3_df}
\alias{.get_Nage_3_df}
\title{Extract Numbers at age 3}
\usage{
.get_Nage_3_df(M)
}
\arguments{
\item{M}{list object created by read_admb function}
}
\value{
dataframe of age3+ biomass
}
\description{
Extract Numbers at age 3
}
