% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-recruitment.R
\name{.get_recruitment_df}
\alias{.get_recruitment_df}
\title{Get recruitment data}
\usage{
.get_recruitment_df(M)
}
\arguments{
\item{M}{list object(s) created by read_admb function}
}
\value{
dataframe of recruitment
}
\description{
Extracts predicted recruitment and approximate asymptotic error-bars
}
\author{
SJD Martell, DN Webber
}
