% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ser.R
\name{.get_ser_df}
\alias{.get_ser_df}
\title{Extract spawning exploitation rate}
\usage{
.get_ser_df(M)
}
\arguments{
\item{M}{list object created by read_admb function}
}
\value{
dataframe of spawning biomass
}
\description{
Extract spawning exploitation rate
}
