% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-srr.R
\name{.get_srr_df}
\alias{.get_srr_df}
\title{Extract stock recruitment results}
\usage{
.get_srr_df(M)
}
\arguments{
\item{M}{list object created by read_admb function}
}
\value{
dataframe of expected and observed SRR
}
\description{
Extract stock recruitment results
}
