% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ssb-rel.R
\name{.get_ssb_rel_df}
\alias{.get_ssb_rel_df}
\title{Extract spawning stock biomass (ssb)}
\usage{
.get_ssb_rel_df(M)
}
\arguments{
\item{M}{list object created by read_admb function}
}
\value{
dataframe of spawning biomass
}
\description{
Spawning biomass may be defined as all males or some combination of males and
females
}
