% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{.run_mod_windows}
\alias{.run_mod_windows}
\title{run pollock model on windows}
\usage{
.run_mod_windows(moddir = mod_dir)
}
\arguments{
\item{moddir}{directory where pollock model should be run}
}
\description{
run pollock model on windows
}
