% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-proj.R
\name{.run_proj}
\alias{.run_proj}
\title{run projection model DOESN"T WORK...dunno why}
\usage{
.run_proj(moddir = mod_dir)
}
\arguments{
\item{moddir}{directory where pollock model should be run}
}
\description{
run projection model DOESN"T WORK...dunno why
}
