% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-proj.R
\name{.run_proj_windows}
\alias{.run_proj_windows}
\title{run projection model for windows}
\usage{
.run_proj_windows(moddir = mod_dir)
}
\arguments{
\item{moddir}{directory where pollock model should be run}
}
\description{
run projection model for windows
}
