% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebswp-package.R
\docType{package}
\name{ebswp-package}
\alias{ebswp}
\alias{ebswp-package}
\title{ebswp: Facilitates Model Running for EBS Pollock}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Facilitates Model Running for EBS Pollock
}
\seealso{
Useful links:
\itemize{
  \item \url{http://afsc-assessments.github.io/EBSpollock/}
  \item \url{http://afsc-assessments.github.io/EBSwp/}
  \item \url{http://afsc-assessments.github.io/ebswp/}
}

}
\author{
\strong{Maintainer}: Jim Ianelli \email{Jim.Ianelli@noaa.gov} (\href{https://orcid.org/0000-0002-7170-8677}{ORCID})

}
\keyword{internal}
