% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tier3-res.R
\name{get_tier3_res}
\alias{get_tier3_res}
\title{Extract Tier 3 Results}
\usage{
get_tier3_res(proj_file)
}
\arguments{
\item{proj_file}{A file path to the projection file.}
}
\value{
A list or vector of results relevant to Tier 3 analyses.
}
\description{
Extracts results relevant to Tier 3 analyses.
}
\examples{
\dontrun{
proj_file_path <- "path/to/proj.csv"
tier3_results <- get_tier3_res(proj_file_path)
}
}
