% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-vars.R
\name{get_vars}
\alias{get_vars}
\title{Extract and Compute Several Metrics from Model Output}
\usage{
get_vars(M, proj_file = NULL)
}
\arguments{
\item{M}{A list or data structure that holds the model output/results. Expected to have several
named elements, including \code{fit}, \code{R}, \code{SSB}, etc.}

\item{proj_file}{(Optional) A file path to a projection file for Tier 3 results.
Default is NULL.}
}
\value{
A list \code{B} that contains various extracted and computed metrics.
}
\description{
This function processes the output of a certain model (possibly related to fisheries)
to extract, compute, and format various metrics.
}
\examples{
\dontrun{
model_result <- list(...) # Example model result here
metrics <- get_vars(model_result)
}
}
