% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-agefit.R
\name{plot_agefit}
\alias{plot_agefit}
\title{Age Composition Data Plotter
t(t()
This function generates a series of plots visualizing the age composition data
from fisheries assessment. It compares observed and predicted data across different years.}
\usage{
plot_agefit(
  x,
  case_label = "2021 assessment",
  gear = "bts",
  type = "survey",
  styr = NULL,
  ageplus = NULL
)
}
\arguments{
\item{case_label}{Label for the specific case or dataset being visualized. Default is "2011 assessment".}

\item{dat}{A data frame, default is \code{mod1}. The structure must contain age composition data, and the names used to create the temporary variables \code{tmp1}, \code{tmp2}, and \code{tmp3}.}

\item{main}{Title of the main plot. Default is an empty string.}

\item{f}{An integer representing the fishery index. Default is 1.}

\item{rec_age}{An integer denoting the recruitment age. Default is 2.}
}
\value{
A multi-panel plot where each panel visualizes age composition data for a
specific year.
}
\description{
Age Composition Data Plotter
t(t()
This function generates a series of plots visualizing the age composition data
from fisheries assessment. It compares observed and predicted data across different years.
}
\details{
The function plots observed fishery age composition data using a bar plot,
overlaying predicted data with points. It organizes the plots in a multi-panel figure,
with each panel representing a year. Age classes are represented on the x-axis.
}
\note{
The function uses the \code{rainbow} color palette, and the colors loop around for
each cohort. The function expects certain naming conventions in the input dataset \code{dat}.
}
\examples{
# Example usage (ensure you have an appropriate dataset 'mod1'):
# AgeFits(dat=mod1, main="Fishery Age Composition", case_label="Sample Assessment")

}
