% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ats.R
\name{plot_ats}
\alias{plot_ats}
\title{Plot Acoustic Trawl Survey Biomass Data}
\usage{
plot_ats(
  M,
  xlab = "Year",
  ylab = "Acoustic trawl survey biomass",
  xlim = NULL,
  ylim = NULL,
  alpha = 0.1,
  biomass = TRUE,
  color = "red"
)
}
\arguments{
\item{M}{List of  model outputs. A list object like the one accepted
by \code{.get_ats_df}.}

\item{xlab}{Label for the x-axis. Default is "Year".}

\item{ylab}{Label for the y-axis. Default is "Acoustic trawl survey biomass".}

\item{xlim}{Optional range for the x-axis.}

\item{ylim}{Optional range for the y-axis.}

\item{alpha}{Opacity for the ribbon, if used. Default is 0.1.}

\item{biomass}{Logical flag. If \code{TRUE}, plots data related to biomass;
if \code{FALSE}, plots data related to another metric. Default is \code{TRUE}.}

\item{color}{Color for the data points. Default is "red".}
}
\value{
A ggplot2 plot object visualizing the acoustic trawl survey data.
}
\description{
This function visualizes the acoustic trawl survey data, either biomass or
another metric, depending on the choice. It provides options for customization
of the resulting plot.
}
