% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-avo.R
\name{plot_avo}
\alias{plot_avo}
\title{Plot avo data}
\usage{
plot_avo(
  M,
  xlab = "Year",
  ylab = "Acoustic return (Sa from AVO) ",
  ylim = NULL
)
}
\arguments{
\item{M}{A list of models where each model has the needed attributes.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{ylim}{Limits for the y-axis.}

\item{xlim}{Limits for the x-axis.}

\item{color}{Color for the points.}

\item{biomass}{Logical indicating if biomass data should be used. If FALSE, other measurements are used.}
}
\value{
A ggplot object with the plotted data.
}
\description{
This function takes a list of models and plots the observed, predicted and confidence intervals for the bottom trawl survey data.
}
