% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-likel.R
\name{plot_likes}
\alias{plot_likes}
\title{Plot Likelihood Values}
\usage{
plot_likes(
  M,
  xlab = "Natural Mortality",
  ylab = "relative -ln Likelihood",
  ylim = NULL,
  xlim = NULL,
  alpha = 0.1,
  legend = TRUE
)
}
\arguments{
\item{M}{A list of models from which likelihood values will be extracted.}

\item{xlab}{Label for the x-axis. Default is "Natural Mortality".}

\item{ylab}{Label for the y-axis. Default is "relative -ln Likelihood".}

\item{ylim}{Limits for the y-axis. Default is NULL.}

\item{xlim}{Limits for the x-axis. Default is NULL.}

\item{alpha}{Alpha transparency level for the plotted lines. Default is 0.1.}

\item{legend}{Logical indicating whether to display a legend. Default is TRUE.}
}
\value{
A ggplot object displaying the negative log likelihood values.
}
\description{
This function plots the negative log likelihood values for different
components against a specified predictor (e.g., natural mortality).
The function uses \code{ggplot2} for visualization.
}
\examples{
# Assuming 'model_list' is a list of models with the appropriate structure
# plot_likes(model_list)

}
