% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mnage.R
\name{plot_mnage}
\alias{plot_mnage}
\title{Plot predicted mean age by gear type}
\usage{
plot_mnage(M, xlab = "Year", ylab = "Mean age", xlim = NULL, ylim = NULL)
}
\arguments{
\item{M}{List object(s) created by read_admb function}

\item{xlab}{the x-label of the figure}

\item{ylab}{the y-label of the figure}

\item{xlim}{is the year range to plot of the figure}

\item{ylim}{is the upper limit of the figure}
}
\value{
Plot of model estimates of mean age against observed (and implied confidence bounds)
}
\description{
Plot predicted mean age by gear type
}
