% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sel.R
\name{plot_sel}
\alias{plot_sel}
\title{Plot selectivity}
\usage{
plot_sel(
  Year = M$Yr,
  sel = M$sel_fsh,
  styr = 1977,
  fage = NULL,
  lage = NULL,
  alpha = 0.2,
  scale = 3.8,
  fill = "purple"
)
}
\arguments{
\item{alpha}{the opacity of the ribbon}

\item{M}{List object(s) created by read_admb function}

\item{xlab}{the x-label of the figure}

\item{ylab}{the y-label of the figure}

\item{ylim}{is the upper limit of the figure}
}
\value{
Plot of model estimates of spawning stock biomass
}
\description{
Plot selectivity
}
