% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-srr.R
\name{plot_srr}
\alias{plot_srr}
\title{Plot Stock-Recruitment Relationship (SRR)}
\usage{
plot_srr(
  M,
  ylab = "Recruits (age 1, millions)",
  xlab = "Female spawning biomass (kt)",
  ylim = NULL,
  xlim = NULL,
  alpha = 0.05,
  ebar = "FALSE",
  leglabs = NULL,
  coverlap = FALSE,
  sizein = 3,
  sizeout = 2,
  yrsin = 1977:2019
)
}
\arguments{
\item{M}{A list or data structure containing model results.}

\item{ylab}{Label for the y-axis. Default is "Recruits (age 1, millions)".}

\item{xlab}{Label for the x-axis. Default is "Female spawning biomass (kt)".}

\item{ylim}{Limits for the y-axis. Default is \code{NULL}.}

\item{xlim}{Limits for the x-axis. Default is \code{NULL}.}

\item{alpha}{Alpha for the ribbons indicating uncertainty. Default is 0.05.}

\item{ebar}{Logical, if \code{TRUE} error bars will be plotted. Default is \code{FALSE}.}

\item{leglabs}{Custom labels for the legend. Default is \code{NULL}.}

\item{coverlap}{Logical, if \code{TRUE} overlapping text labels will be checked and prevented. Default is \code{FALSE}.}

\item{sizein}{Font size for the in-sample text labels. Default is 3.}

\item{sizeout}{Font size for the out-of-sample text labels. Default is 2.}

\item{yrsin}{Years to be considered for in-sample. Default is 1977:2019.}
}
\value{
A ggplot object containing the SRR plot.
}
\description{
This function plots the stock-recruitment relationship (SRR) using ggplot2.
It can handle multiple models and provides flexibility in display options.
}
\examples{
# Assuming 'model_list' contains the relevant model results:
# plot_srr(M = model_list, ylim = c(0, 1e6), xlim = c(0, 2000))

}
