% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ssb.R
\name{plot_ssb}
\alias{plot_ssb}
\title{Plot predicted spawning stock biomass (ssb)}
\usage{
plot_ssb(
  M,
  xlab = "Year",
  ylab = "Female spawning biomass (kt)",
  ylim = NULL,
  xlim = NULL,
  breaks = seq(1990, 2022, 2),
  alpha = 0.1,
  legend = TRUE,
  order = NULL
)
}
\arguments{
\item{M}{List object(s) created by read_admb function}

\item{xlab}{the x-label of the figure}

\item{ylab}{the y-label of the figure}

\item{ylim}{is the upper limit of the figure}

\item{alpha}{the opacity of the ribbon}
}
\value{
Plot of model estimates of spawning stock biomass
}
\description{
Spawning biomass may be defined as all males or some combination of males and
females
}
