% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ssb-rel.R
\name{plot_ssb_rel}
\alias{plot_ssb_rel}
\title{Plot predicted spawning stock biomass (ssb)}
\usage{
plot_ssb_rel(
  M,
  xlab = "Year",
  ylab = "Relative female spawning biomass",
  ylim = NULL,
  xlim = NULL,
  legend = TRUE,
  breaks = seq(1990, 2022, 5),
  alpha = 0.1
)
}
\arguments{
\item{M}{List object(s) created by read_admb function}

\item{xlab}{the x-label of the figure}

\item{ylab}{the y-label of the figure}

\item{ylim}{is the upper limit of the figure}

\item{alpha}{the opacity of the ribbon}
}
\value{
Plot of model estimates of spawning stock biomass
}
\description{
Spawning biomass may be defined as all males or some combination of males and
females
}
