% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_Tier3_tbl.R
\name{print_Tier3_tbl}
\alias{print_Tier3_tbl}
\title{Generate and Print Tier 3 Summary Tables}
\usage{
print_Tier3_tbl(mod_number, run = FALSE)
}
\arguments{
\item{mod_number}{Integer. Specifies the model configuration number.}

\item{run}{Logical. If \code{TRUE}, runs the projection model for the specified configuration. Default is \code{FALSE}.}
}
\value{
Prints formatted tables for Catch, SSB, F, and ABC. Does not return a value.
}
\description{
This function generates and prints summary tables for a specified model configuration.
It calculates mean values for various metrics across different scenarios and displays
formatted tables for Catch, Spawning Stock Biomass (SSB), Fishing Mortality (F), and Allowable Biological Catch (ABC).
}
\details{
The function reads projection results from a CSV file, groups data by alternative and year,
calculates mean values for specific metrics, and formats these into tables for each metric.
The tables are printed with captions based on the model configuration name.
}
\examples{
# Example usage
print_Tier3_tbl(mod_number = 1, run = TRUE)

}
