% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-admb.R
\name{read_dat}
\alias{read_dat}
\title{Read ADMB Data File}
\usage{
read_dat(fn)
}
\arguments{
\item{fn}{Character. The name of the ADMB data file.}
}
\value{
A list containing various ADMB outputs from the data file.
}
\description{
Reads an ADMB data file and returns a list of outputs in R.
}
