% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab-fit.R
\name{tab_fit}
\alias{tab_fit}
\title{Tabulate fits from models using extended metrics.}
\usage{
tab_fit(M, mod_scen = NULL)
}
\arguments{
\item{M}{List containing model outputs.}

\item{mod_scen}{Optional vector of integers indicating which models in M to use. Default is all models.}
}
\value{
Returns a dataframe with extended fit metrics.
}
\description{
This function takes a list of model outputs and compiles a dataframe with a broader set of fit metrics.
}
