% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab-fit.R
\name{upperCaseFirst}
\alias{upperCaseFirst}
\title{Helper function to convert the first character of a string to uppercase.}
\usage{
upperCaseFirst(s)
}
\arguments{
\item{s}{The input string.}
}
\value{
Returns the input string with the first character capitalized.
}
\description{
Helper function to convert the first character of a string to uppercase.
}
