% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-dat.R
\name{write_dat}
\alias{write_dat}
\title{Write Data to Text File}
\usage{
write_dat(output_file = "runs/dat/newfile.dat", indata = in_data)
}
\arguments{
\item{output_file}{A character string specifying the name of the output text file.
Default is "output.txt".}

\item{indata}{A list containing the data elements to be written to the file.
Each element can be a matrix, a vector, or any other value. Default is \code{in_data}.}
}
\value{
This function doesn't return a value; it writes to the specified output file.
}
\description{
This function takes a list of data elements (matrices, vectors, or other values)
and writes them to a specified text file. Each data element in the list is preceded
by a comment line indicating its name.
}
\examples{
# Assuming 'data_list' is a list of data elements
# write_dat(output_file = "sample_output.txt", indata = data_list)
}
